requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
numpy
torch
torchvision
lightning
matplotlib
pandas
scikit-learn
pillow
pyyaml
scikit-image
matplotlib
seaborn
captum
opencv-python
